+++ src/main/java/org/scoooting/dto/EndRentalRequest.java
@@ -2,4 +2,4 @@ package org.scoooting.dto;
 
 import jakarta.validation.constraints.NotNull;
 
-public record EndRentalRequest(@NotNull Double endLatitude, @NotNull Double endLongitude) {}
+public record EndRentalRequest(@NotNull Float endLatitude, @NotNull Float endLongitude) {}
