+++ src/main/java/org/scoooting/entities/Scooter.java
@@ -3,12 +3,10 @@ package org.scoooting.entities;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.scoooting.entities.enums.ScootersStatus;
 import org.springframework.data.annotation.Id;
-import org.springframework.data.annotation.Transient;
 import org.springframework.data.relational.core.mapping.Table;
 
 @Table("scooters")
@@ -28,7 +26,7 @@ public class Scooter {
     @Size(max = 32)
     private ScootersStatus status;
 
-    private Double latitude;
-    private Double longitude;
+    private Float latitude;
+    private Float longitude;
 
 }
